requirements.txt
aiogram==3.22.0
python-dotenv==1.2.1
fluent-runtime==0.4.0
aiogram_i18n==1.4
fluent_compiler==1.1